+++ src/main/java/net/adeptropolis/frogspawn/graphs/functions/GraphFunction.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) Florian Schaefer 2020.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package net.adeptropolis.frogspawn.graphs.functions;
+
+import net.adeptropolis.frogspawn.graphs.Graph;
+
+@FunctionalInterface
+interface GraphFunction<T> {
+
+  T apply(Graph graph);
+
+}